Fixed wheel speeds for large rover speeds.
Previously when the rover's speed exceeded the maximum speed of one wheel, it wouldn't recalculate the other whhel speeds to account for one of the wheels being limited.

Rover speed is now limited so that it can't exceed a speed that will cause one wheel to exceed the largest velocity.

Fixed a bug with wheel angle and tidied up code a bit.